move workfiles interface to separate file
